backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlmodel==0.0.24
psycopg[binary]==3.2.9
alembic==1.16.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==6.4.0
httpx==0.28.1
structlog==25.4.0
email-validator>=2.0.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-multipart==0.0.9
